Refactor package: Create separate GUI package
