+++ src/main/java/duke/gui/Launcher.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.gui;
 
 import javafx.application.Application;
